+++ src/errors.ts
@@ -38,7 +38,7 @@ export type FileError = {
   /**
    * Error code.
    */
-  readonly code: 'NO_JSON_FILE_FOUND' | 'UNREFERENCED_JSON_FILE'
+  readonly code: 'NO_JSON_FILE_FOUND' | 'UNREFERENCED_JSON_FILE' | 'CIRCULAR_REFERENCE'
   /**
    * Error message.
    */
@@ -53,45 +53,4 @@ export type FileError = {
   readonly jsonUrl: string
 }
 
-export type UnreferencedJsonError = {
-  /**
-   * Error code.
-   */
-  readonly code: 'UNREFERENCED_JSON_FILE'
-  /**
-   * Error message.
-   */
-  readonly message: string
-  /**
-   * Directory.
-   */
-  readonly readMeUrl: string
-  /**
-   * URL of JSON file.
-   */
-  readonly jsonUrl: string
-}
-
-export type CircularReferenceError = {
-  /**
-   * Error code.
-   */
-  readonly code: 'CIRCULAR_REFERENCE'
-
-  /**
-   * Error message.
-   */
-  readonly message: string
-
-  /**
-   * URL of `readme.md` file.
-   */
-  readonly readMeUrl: string
-
-  /**
-   * URL of JSON file.
-   */
-  readonly jsonUrl: string
-}
-
-export type Error = JsonParseError | FileError | NotAutoRestMarkDown | CircularReferenceError | UnreferencedJsonError
+export type Error = JsonParseError | FileError | NotAutoRestMarkDown