+++ src/test/avocado-test.ts
@@ -187,15 +187,14 @@ describe('avocado', () => {
       {
         code: 'NO_JSON_FILE_FOUND',
         message: 'The JSON file is not found but it is referenced from the readme file.',
-        readMeUrl: '/home/ruowan/work/avocado/src/test/referenced_common_spec/specification/service/readme.md',
-        jsonUrl:
-          '/home/ruowan/work/avocado/src/test/referenced_common_spec/specification/common/specs/no_such_file.json',
+        readMeUrl: path.resolve('src/test/referenced_common_spec/specification/service/readme.md'),
+        jsonUrl: path.resolve('src/test/referenced_common_spec/specification/common/specs/no_such_file.json'),
       },
       {
         code: 'UNREFERENCED_JSON_FILE',
         message: 'The JSON file is not referenced from the readme file.',
-        readMeUrl: '/home/ruowan/work/avocado/src/test/referenced_common_spec/specification/common/readme.md',
-        jsonUrl: '/home/ruowan/work/avocado/src/test/referenced_common_spec/specification/common/specs/orphan.json',
+        readMeUrl: path.resolve('src/test/referenced_common_spec/specification/common/readme.md'),
+        jsonUrl: path.resolve('src/test/referenced_common_spec/specification/common/specs/orphan.json'),
       },
     ] as const
     assert.deepStrictEqual(r, expected)